Feature/1.2 to master

+++ code/software/libs/src/cstdlib/include/ctype.h
@@ -18,7 +18,7 @@
 
 #include <stdint.h>
 
-typedef   uint32_t    size_t;
+typedef unsigned long size_t;
 
 #endif
